+++ tests/deprecation-validator.js
@@ -11,13 +11,11 @@ const titleIdentifier = fs.readFileSync(process.cwd() + '/tests/mocks/schemas/de
 const titleBeforeId = fs.readFileSync(process.cwd() + '/tests/mocks/schemas/deprecated/edit-with-deprecated-title-before-id.yml');
 const titleAfterId = fs.readFileSync(process.cwd() + '/tests/mocks/schemas/deprecated/edit-with-deprecated-title-after-id.yml');
 const titleComponents = fs.readFileSync(process.cwd() + '/tests/mocks/schemas/deprecated/edit-with-deprecated-title-components.yml');
-const staticFilters = fs.readFileSync(process.cwd() + '/tests/mocks/schemas/deprecated/edit-with-deprecated-static-filters.yml');
 
 const deprecatedSchemaIdentifier = ymljs.parse(titleIdentifier.toString());
 const deprecatedSchemaBeforeId = ymljs.parse(titleBeforeId.toString());
 const deprecatedSchemaAfterId = ymljs.parse(titleAfterId.toString());
 const deprecatedSchemaTitleComponents = ymljs.parse(titleComponents.toString());
-const deprecatedSchemaStaticFilters = ymljs.parse(staticFilters.toString());
 
 describe('deprecation-validator', () => {
 
@@ -72,13 +70,5 @@ describe('deprecation-validator', () => {
 		assert(executeSpy.calledOnce);
 	});
 
-	it('should find a deprecated property staticFilters', async () => {
-
-		const executeSpy = sinon.spy(deprecationValidator, 'execute');
-
-		deprecationValidator.execute(deprecatedSchemaStaticFilters);
-
-		assert(executeSpy.calledOnce);
-	});
 
 });