+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- staticFilter deprecated
+
 ## [2.124.0] - 2024-03-14
 
 ### Changed